modifié :         __pycache__/config.cpython-35.pyc 	modifié :         action.py 	modifié :         actions.cfg 	modifié :         carte.py 	modifié :         chemin.py 	modifié :         communication.py 	modifié :         config.py 	modifié :         deplacement.py 	modifié :         fonction_chemin_grossier.py 	modifié :         main.py 	modifié :         parcours.py 	modifié :         robot.py

+++ config.py
@@ -4,7 +4,6 @@ import os
 
 class Config():
    
-
    fichier = open("config.cfg", "r")
    ligne = fichier.read().split("\n")
    for i in range(len(ligne)-1):
@@ -45,7 +44,7 @@ class Config():
          except AssertionError:
             print("Le fichier de configuration a ete mal encodé")
    dx = longueur_terrain/X
-   Y = largeur_terrain/dx
+   Y = int(largeur_terrain/dx)
    
 
    def __init__(self):
